Commenting out unclear lines
There are some problematic lines in here.  Some did not match their intended targets due to incomplete regex syntax that would miss leading or trailing strings, some are repeats, and others look out of place.  I think I'm going to add a flag and bits to tally how many times each pattern was matched with debug on the user agent that matched, along with a report on all user agents that made it past.